+++ packages/client/src/types.ts
@@ -9,6 +9,7 @@ import type {
   MemberResponse,
   OwnCapability,
   ReactionResponse,
+  AudioSettingsRequestDefaultDeviceEnum,
   StartRecordingRequest,
   StartRecordingResponse,
 } from './gen/coordinator';
@@ -343,6 +344,33 @@ export type StartCallRecordingFnType = {
     type: CallRecordingType,
   ): Promise<StartRecordingResponse>;
 };
+
+export type StreamRNVideoSDKGlobals = {
+  callManager: {
+    /**
+     * Sets up the in call manager.
+     */
+    setup({
+      default_device,
+    }: {
+      default_device: AudioSettingsRequestDefaultDeviceEnum;
+    }): void;
+
+    /**
+     * Starts the in call manager.
+     */
+    start(): void;
+
+    /**
+     * Stops the in call manager.
+     */
+    stop(): void;
+  };
+};
+
+declare global {
+  var streamRNVideoSDK: StreamRNVideoSDKGlobals | undefined;
+}
 
 /**
  * The options to pass to {@link Call.join} method.